requirements.txt
alto-xml==0.0.5
certifi==2023.5.7
charset-normalizer==3.2.0
Cython==3.0.0
easyocr==1.7.0
filelock==3.12.2
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.25.1
opencv-python-headless==4.8.0.74
packaging==23.1
pdf2image==1.16.3
Pillow==9.5.0
pyclipper==1.3.0.post4
pytesseract==0.3.10
python-bidi==0.4.2
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.1
shapely==2.0.1
six==1.16.0
svgwrite==1.4.3
sympy==1.12
tesseract-ocr-data==1.6
tesseract-ocr-utils==0.0.6
tifffile==2023.7.18
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
